댓글 crud기능 4차수정

+++ src/main/java/com/sparta/travelnewsfeed/service/CommentService.java
@@ -7,14 +7,8 @@ import com.sparta.travelnewsfeed.entity.Comment;
 import com.sparta.travelnewsfeed.entity.Post;
 import com.sparta.travelnewsfeed.repository.CommentRepository;
 import com.sparta.travelnewsfeed.repository.PostRepository;
-import com.sparta.travelnewsfeed.repository.UserRepository;
 import com.sparta.travelnewsfeed.user.User;
-import com.sparta.travelnewsfeed.user.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -31,33 +25,18 @@ public class CommentService {
 
     private final PostRepository postRepository;
 
-    private final UserRepository userRepository;
-
     private final PasswordEncoder passwordEncoder;
 
     @Transactional
-    public CommentResponseDto createComment(CommentCreateRequestDto dto) {
-        String username = SecurityContextHolder.getContext().getAuthentication().getName();
-        Post post = postRepository.findById(dto.getPostId())
-                .orElseThrow(() -> new IllegalArgumentException("게시물 을 찾을 수 없습니다. id: " + dto.getPostId()));
-        User user = userRepository.findByUsername(dto.getUsername())
-                .orElseThrow(() -> new IllegalArgumentException("유저를 찾을 수 없습니다. 이름: " + dto.getUsername()));
+    public CommentResponseDto createComment(Long postId, User user, CommentCreateRequestDto dto) {
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new IllegalArgumentException("게시물을 찾을 수 없습니다."));
+        Comment comment = new Comment(post, user, dto);
+        commentRepository.save(comment);
 
-        Comment comment = new Comment();
-        comment.setText(dto.getText());
-        comment.setPost(post);
-        comment.setUser(user);
-
-        Comment savedComment = commentRepository.save(comment);
-
-        return new CommentResponseDto(
-                savedComment.getId(),
-                savedComment.getText(),
-                savedComment.getCreatedAt(),
-                savedComment.getUpdatedAt(),
-                user.getUsername()
-        );
+        return new CommentResponseDto(comment);
     }
+
 
     public CommentResponseDto getCommentById(Long id) {
         Comment comment = commentRepository.findById(id)
@@ -74,9 +53,7 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentResponseDto updateComment(Long id, CommentUpdateRequestDto dto) {
-        User user = userRepository.findByUsername(dto.getUsername())
-                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다.: " + dto.getUsername()));
+    public CommentResponseDto updateComment(User user, Long id, CommentUpdateRequestDto dto) {
 
         if (!passwordEncoder.matches(dto.getPassword(), user.getPassword())) {
             throw new IllegalArgumentException("잘못된 비밀번호");
@@ -85,8 +62,8 @@ public class CommentService {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("id를 찾을 수 없습니다. id: " + id));
 
-        if (!comment.getUser().equals(user)) {
-            throw new SecurityException("자신의 댓글만 수정할 수 있습니다.1");
+        if (!comment.getUser().getId().equals(user.getId())) {
+            throw new SecurityException("자신의 댓글만 수정할 수 있습니다.");
         }
 
         comment.setText(dto.getText());
@@ -96,9 +73,14 @@ public class CommentService {
                 updatedComment.getCreatedAt(), updatedComment.getUpdatedAt(), user.getUsername());
     }
 
-    public void deleteComment(Long id) {
+    public void deleteComment(User user, Long id) {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("댓글 id를 찾을 수 없습니다. id: " + id));
-        commentRepository.delete(comment);
+
+        if (comment.getUser().getUsername().equals(user.getUsername())) {
+            commentRepository.delete(comment);
+        } else {
+            throw new SecurityException("자신의 댓글만 삭제할 수 있습니다.");
+        }
     }
 }
